requirements.txt
cohere==4.37
streamlit==1.28.0
python-dotenv==1.0.0 